neural-search/requirements.txt
# Core dependencies
numpy>=1.24.0
scipy>=1.10.0

# Vector search
scikit-learn>=1.3.0

# API framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
python-multipart>=0.0.6

# Rate limiting
slowapi>=0.1.9

# GraphQL
strawberry-graphql[fastapi]>=0.213.0

# Embeddings and NLP
transformers>=4.35.0
sentence-transformers>=2.2.2
torch>=2.1.0
tokenizers>=0.15.0

# Text processing
regex>=2023.10.3

# Storage and persistence
h5py>=3.10.0
redis>=5.0.0

# Async support
aiohttp>=3.9.0
asyncio>=3.4.3

# Monitoring and logging
prometheus-client>=0.19.0

# Utils
python-dotenv>=1.0.0
click>=8.1.7

# Testing (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

# Development (optional)
black>=23.0.0
flake8>=6.1.0
mypy>=1.7.0